+++ transports/small-webrtc-transport/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to **Pipecat Small WebRTC Transport** will be documented in 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Implemented support for passing custom data and headers to the webrtc connection (offer/answer) endpoint. The `webrtcUrl` can now be a `string` or a `APIRequest` type. Server endpoints can access custom `requestData` under that same field.
+
 ## [1.3.0]
 
 ### Added